feat(api): expose bitcoin types for the API
In some cases, you need to use the bitcoin primitive to convert a
type into another format or build a type from a hex string.

This commit will reexport the primitive used in the client without
forcing the caller to add the bitcoin crate as direct dependency.

The motivation to have this is when you are interacting with a system
that is not rust, and you do not need the rust-bitcoin.

Signed-off-by: Vincenzo Palazzo <vincenzopalazzodev@gmail.com>